feat: add function toDoSubmit to add-todo.js
Add function toDoSubmit to module add-todo.js that adds a new toDo with
the contents of the input field as the name of the title. The function
ensures that the new toDo is added to the correct category. The function
then execures displayHub(), which will display the new toDo item.

Add event listener to add new todo form that executes function
toDoSubmit upon form submission.