fix: prevent isComplete() from throwing errors
Add a return to function isComplete in module todo-functions.js to
stop the loop once the obj is found and operated on.

Add operational chaining to isComplete to safely access nested
properties of an object without causing a runtime error if any part of
the chain is null or undefined.